Add query summary matching
This has been copied from the openfoodnetwork repository and adapted
slightly. It also needed one little fix.

This approach is quite different to our previous one and I'm wondering
if we use `User` and `Create` instead of `users` and `insert` to stay on
the higher level of Active Record. But for now this is can be a simple
replacement for the openfoodnetwork spec helper and we can evolve it
from there.